Added 32-bit fixes from QDP-JIT issue #22

+++ lib/qdp_random.cc
@@ -26,10 +26,10 @@ namespace RNG
   LatticeSeed *lattice_ran_mult;
 
     //! Find the number of bits required to represent x.
-  int numbits(int x)
+  int numbits(size_t x)
   {
     int num = 1;
-    int iceiling = 2;
+    size_t iceiling = 2;
     while (iceiling <= x)
     {
       num++;
@@ -133,7 +133,7 @@ namespace RNG
     Seed aa;
     Seed aamult;
 
-    int ibit = Layout::vol();
+    size_t ibit = Layout::vol();
     aa = ran_mult;
 //    ran_mult_n = 1;    // produces def    ran_mult_n = ran_mult^{vol}
     ran_mult_n = ran_mult;   // produces def  ran_mult_n = ran_mult^{vol+1}
